requirements.txt
mysql-connector-python
requests
python-dotenv
google-api-python-client
pygsheets